Move VerificationModal to Authentication
