+++ front/app/containers/Authentication/VerificationModal/VerificationError.tsx
