+++ front/app/containers/Authentication/VerificationModal/VerificationMethodButton.tsx
