+++ front/app/containers/Authentication/VerificationModal/VerificationSuccess.tsx
