+++ front/app/containers/Authentication/VerificationModal/index.tsx
