+++ front/app/containers/Authentication/VerificationModal/isLastVerificationMethod.ts
