+++ front/app/containers/Authentication/VerificationModal/messages.ts
