+++ front/app/containers/Authentication/VerificationModal/verificationModalEvents.ts
